refactor(games): extract StatementGameService base for true/false submit ss-148
